Add personal allowance logic
