--- openfisca_uk/tests/policy/baseline/finance/benefit/family/housing_benefit.yaml
@@ -1,52 +1,0 @@
-- name: Housing Benefit for a single person with small earnings
-  period: 2021
-  absolute_error_margin: 5
-  input:
-    age: 26
-    weekly_hours: 31
-    employment_income: 0
-    tenure_type: RENT_FROM_COUNCIL
-    would_claim_hb: true
-    claims_WTC: true
-    benunit_rent: 88 * 52
-    is_household_head: true
-  output:
-    WTC_basic_element: 3040
-    WTC_worker_element: 825
-    working_tax_credit: 3040 + 825
-    housing_benefit: 88 * 52
-- name: Check no change up until the phase-out point
-  period: 2021
-  absolute_error_margin: 5
-  input:
-    age: 26
-    weekly_hours: 31
-    employment_income: 3801 - (3865 - 52 * (5 + 37.1))
-    tenure_type: RENT_FROM_COUNCIL
-    would_claim_hb: true
-    claims_WTC: true
-    benunit_rent: 88 * 52
-    is_household_head: true
-  output:
-    WTC_basic_element: 3040
-    WTC_worker_element: 825
-    working_tax_credit: 3865
-    housing_benefit_applicable_income: 3801
-    housing_benefit: 88 * 52
-- name: Check phase-out rate is 65%
-  period: 2021
-  absolute_error_margin: 5
-  input:
-    age: 26
-    weekly_hours: 31
-    employment_income: 3801 - (3865 - 52 * (5 + 37.1)) + 100
-    tenure_type: RENT_FROM_COUNCIL
-    would_claim_hb: true
-    claims_WTC: true
-    benunit_rent: 88 * 52
-    is_household_head: true
-  output:
-    WTC_basic_element: 3040
-    WTC_worker_element: 825
-    working_tax_credit: 3865
-    housing_benefit: 88 * 52 - 65
